requirements.txt
pygame==2.6.1
pillow==10.3.0
psutil==5.9.0

